requirements.txt
django
django-allauth
google-api-python-client
google-auth
